refactor(observability): 分离体验复核清单
